refactor(client): co-locate HeroBootAnimation, HeroSystemFailure, RoleTyper with unit tests
