+++ components/client/HeroBootAnimation/index.ts
@@ -1,0 +1,1 @@
+export { HeroBootAnimation } from './HeroBootAnimation';
