+++ components/client/HeroSystemFailure/index.ts
@@ -1,0 +1,1 @@
+export { HeroSystemFailure } from './HeroSystemFailure';
