+++ components/client/RoleTyper/index.ts
@@ -1,0 +1,1 @@
+export { RoleTyper } from './RoleTyper';
